+++ .release-it.cjs
@@ -24,7 +24,7 @@ module.exports = {
 			'node bin/release-changelog.mjs promote ${version}',
 			'node bin/release-readme.mjs update ${version}',
 			'npm run build',
-			'git add package.json package-lock.json packages/core/package.json packages/audio/package.json packages/midi/package.json packages/jukette/package.json CHANGELOG.md README.md packages/core/dist packages/audio/dist packages/midi/dist packages/jukette/dist',
+			'git add package.json package-lock.json apps/docs/package.json packages/core/package.json packages/audio/package.json packages/midi/package.json packages/soundcloud/package.json packages/jukette/package.json CHANGELOG.md README.md packages/core/dist packages/audio/dist packages/midi/dist packages/soundcloud/dist packages/jukette/dist',
 		],
 		'before:release': 'npm run publish:workspaces:dry-run',
 		'before:github:release': 'npm run publish:workspaces',